Start on interpreter
The environment is empty so it's sorta useless.

I have low confidence that I have the right idea about references
and borrowing here. Here's my thinking:

When we pass an object to the evaluator, we're not really making a
new reference to it: the evaluator simply borrows our reference.
In GC terms, as long as the original Rc is alive somewhere on the
stack, Rust isn't going to clean it up even if eval is only
borrowing that Rc.
The evaluator however _returns_ a new reference, not a borrowed
reference. This is necessary since the evaluator can return new
objects, but also makes sense if it returns a given object, e.g.
from a self-evaluating object. We've created, if only temporarily,
a new way to get at that object, independent of however we
originally referenced it.
And in particular, Cons (and other data structures) have an actual
Rc, not a borrow. In GC terms, cons has actual pointers and can
keep stuff alive (as long as it is alive).
Boy I hope I'm not totally off base here.