Start populating ground environment
This setup is... inelegant, to say the least. But it works.

Using "raw" function pointers as I do here instead of closures is
not convenient, but closures have other problems. They are typed
as traits, which means they have to be put in a dynamically typed
Box or the like. That in turn prevents Debug and Eq from being
automatically derived for anything containing them. And even with
that out of the way I was getting inscrutable errors about
lifetimes. So: function pointers it is.