Use GC library
I really wanted to make my own GC, but I don't think I'm advanced
enough to figure it out, and I don't think it's possible in
non-experimental Rust. For example this library needs the
std::marker::Unsize trait, which is currently experimental.
Something for later. On the bright side, this was really easy to
set up.

This is just a dumb mark-sweep (which is what I was planning on
starting with anyway). Hopefully I'll be able to make my own GC
later, if only for my own edification.