Still figuring out borrowing
I think *foo here is a place expression, meaning that match will
refer to the actual object by reference, rather than make a temp.
That in turn means the fields need to be referred to by reference,
which is what we want anyway. Conceptually assoc/whatever should
not be copying any conses, even just in the stack.